xdg-mime: do not report multiple desktop files (BR60329)
since commit 050333e3, xdg-mime checks both defaults.list and
mimeinfo.cache. When a result is present in both files, it returns both
results separated by a space.

When this is the case, the desktop file name is made of the duplicated
filename and the following chain of executions does not work properly.

This commit tries to select the first found line before extracting the
desktop file name.

Signed-off-by: Lionel Orry <lionel.orry@gmail.com>