xdg-open: Fixes LXQt behavior
Commit 6387086e4938d568c2bab185632f60e1619b3f68 introduced LXQt support.
In xdg-open it assumed that LXQt and LXDE are the same. They aren't. LXQt
does not have pcmanfm, it has pcmanfm-qt.
LXQt doesn't want to rely on pcmamfm-qt to handle it: LXQt is very modular
and can be deployed without pcmanfm-qt.

open_generic() works for LXQt but it depends on mimeopen and mimetype. In
the future we will want, for sure, to drop those dependencies. That's the
reason for the existence of open_lxqt().