xdg-screensaver: fix the freedesktop screensaver
This was not correctly implemented, and perhaps no-one noticed because until
now it was only used for KDE 3. But now GNOME 3 needs it!

The main problem is that, like the gnome-screensaver method, inhibiting the
screensaver requires a long-held DBus connection.

So, reuse the embedded Perl script from screensaver_gnome_screensaver,
parametrized suitably, and implement the ‘suspend’ and ‘resume’ commands for
freedesktop in the same way as for gnome-screensaver.

Also rewrite the code for ‘reset’, which was incorrectly calling
‘UnInhibit’, which does not necessarily disable the screensaver (quite the
reverse, possibly!). Instead, call ‘SimulateUserActivity’, as for the
gnome-screensaver case.

Note that on a current GNOME system (e.g. GNOME 42) some of the DBus
freedesktop endpoints are advertised but unimplemented; xdg-screensaver will
give an error for these cases.