Check if defaults file exists during installation
If a desktop file is installed with xdg-desktop-menu install, the script
tries to unconditionally access $XDG_DATA_HOME/applications/defaults.list,
which may be missing. This is the case when installing into a staging
directory during the FreeBSD package building process. In that case, the
call to grep blocks indefinitely, because grep tries to read from stdin
instead. Check for existence of the defaults file before calling grep on
the defaults file.